+++ src/api.js
@@ -1,20 +1,20 @@
 import {
   ApiCaller,
-  DocumentUploader,
-  WalletsManager,
+  DocumentsManager,
   FactorsManager,
+  WalletsManager,
 } from '@tokend/js-sdk'
 
 export const api = new ApiCaller()
-export const documentUploader = new DocumentUploader({
-  apiCaller: api,
-})
 
+export const documentsManager = new DocumentsManager()
 export const walletsManager = new WalletsManager()
 export const factorsManager = new FactorsManager()
 
 export function useWallet (newWallet) {
   api.useWallet(newWallet)
-  walletsManager.useApi(api)
+
+  documentsManager.useApi(api)
   factorsManager.useApi(api)
+  walletsManager.useApi(api)
 }