+++ src/vue/App.vue
@@ -48,7 +48,7 @@ import {
   mapActions,
 } from 'vuex'
 import { Sdk } from '@/sdk'
-import { api, documentUploader, walletsManager, factorsManager } from '@/api'
+import { api, documentsManager, walletsManager, factorsManager } from '@/api'
 import { vuexTypes } from '@/vuex'
 import { Wallet } from '@tokend/js-sdk'
 import { ErrorTracker } from '@/js/helpers/error-tracker'
@@ -103,7 +103,7 @@ export default {
     async initApp () {
       await Sdk.init(config.HORIZON_SERVER)
       api.useBaseURL(config.HORIZON_SERVER)
-      documentUploader.useStorageURL(config.FILE_STORAGE)
+      documentsManager.useStorageURL(config.FILE_STORAGE)
 
       const { data: networkDetails } = await api.getRaw('/')
       api.useNetworkDetails(networkDetails)